Single out GraphQL execution code
Architecture-wise it's not part of the routing
